collapse autofmt crate
